Added BaseWriteContext and BaseReadContext
